refactor bulk actions
