+++ src/pages/Search/SearchPageNarrow.tsx
@@ -4,28 +4,26 @@ import {View} from 'react-native';
 import Animated, {clamp, useAnimatedScrollHandler, useAnimatedStyle, useSharedValue, withTiming} from 'react-native-reanimated';
 import {scheduleOnRN} from 'react-native-worklets';
 import FullPageNotFoundView from '@components/BlockingViews/FullPageNotFoundView';
-import type {DropdownOption} from '@components/ButtonWithDropdownMenu/types';
 import {useFullScreenBlockingViewActions} from '@components/FullScreenBlockingViewContextProvider';
 import HeaderWithBackButton from '@components/HeaderWithBackButton';
-import type {PaymentMethodType} from '@components/KYCWall/types';
 import NavigationTabBar from '@components/Navigation/NavigationTabBar';
 import NAVIGATION_TABS from '@components/Navigation/NavigationTabBar/NAVIGATION_TABS';
 import TopBar from '@components/Navigation/TopBar';
 import ScreenWrapper from '@components/ScreenWrapper';
 import {ScrollOffsetContext} from '@components/ScrollOffsetContextProvider';
 import Search from '@components/Search';
-import {useSearchSelectionContext} from '@components/Search/SearchSelectionContext';
 import SearchPageFooter from '@components/Search/SearchPageFooter';
 import SearchFiltersBar from '@components/Search/SearchPageHeader/SearchFiltersBar';
 import SearchPageHeader from '@components/Search/SearchPageHeader/SearchPageHeader';
 import SearchSelectionBar from '@components/Search/SearchPageHeader/SearchSelectionBar';
-import type {SearchHeaderOptionValue} from '@components/Search/SearchPageHeader/SearchPageHeader';
+import {useSearchSelectionContext} from '@components/Search/SearchSelectionContext';
 import type {BankAccountMenuItem, SearchParams, SearchQueryJSON} from '@components/Search/types';
 import useAndroidBackButtonHandler from '@hooks/useAndroidBackButtonHandler';
 import useLocalize from '@hooks/useLocalize';
 import useNetwork from '@hooks/useNetwork';
 import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import useScrollEventEmitter from '@hooks/useScrollEventEmitter';
+import useSearchBulkActions from '@hooks/useSearchBulkActions';
 import useStyleUtils from '@hooks/useStyleUtils';
 import useThemeStyles from '@hooks/useThemeStyles';
 import useWindowDimensions from '@hooks/useWindowDimensions';
@@ -47,26 +45,15 @@ const ANIMATION_DURATION_IN_MS = 300;
 type SearchPageNarrowProps = {
     queryJSON?: SearchQueryJSON;
     metadata?: SearchResultsInfo;
-    headerButtonsOptions: Array<DropdownOption<SearchHeaderOptionValue>>;
     searchResults?: SearchResults;
     isMobileSelectionModeEnabled: boolean;
     currentSelectedPolicyID?: string | undefined;
     currentSelectedReportID?: string | undefined;
-    confirmPayment?: (paymentType: PaymentMethodType | undefined) => void;
     latestBankItems?: BankAccountMenuItem[] | undefined;
 };
 
-function SearchPageNarrow({
-    queryJSON,
-    headerButtonsOptions,
-    searchResults,
-    isMobileSelectionModeEnabled,
-    metadata,
-    currentSelectedPolicyID,
-    currentSelectedReportID,
-    latestBankItems,
-    confirmPayment,
-}: SearchPageNarrowProps) {
+function SearchPageNarrow({queryJSON, searchResults, isMobileSelectionModeEnabled, metadata, currentSelectedPolicyID, currentSelectedReportID, latestBankItems}: SearchPageNarrowProps) {
+    const {headerButtonsOptions} = useSearchBulkActions();
     const {translate} = useLocalize();
     const {shouldUseNarrowLayout} = useResponsiveLayout();
     const {windowHeight} = useWindowDimensions();
@@ -214,7 +201,6 @@ function SearchPageNarrow({
                                             topBarOffset.set(StyleUtils.searchHeaderDefaultOffset);
                                             setSearchRouterListVisible(true);
                                         }}
-                                        headerButtonsOptions={headerButtonsOptions}
                                         handleSearch={handleSearchAction}
                                         isMobileSelectionModeEnabled={isMobileSelectionModeEnabled}
                                     />
@@ -224,10 +210,8 @@ function SearchPageNarrow({
                                         (isMobileSelectionModeEnabled ? (
                                             <SearchSelectionBar
                                                 queryJSON={queryJSON}
-                                                headerButtonsOptions={headerButtonsOptions}
                                                 currentSelectedPolicyID={currentSelectedPolicyID}
                                                 currentSelectedReportID={currentSelectedReportID}
-                                                confirmPayment={confirmPayment}
                                                 latestBankItems={latestBankItems}
                                             />
                                         ) : (
@@ -253,13 +237,11 @@ function SearchPageNarrow({
                         />
                         <SearchPageHeader
                             queryJSON={queryJSON}
-                            headerButtonsOptions={headerButtonsOptions}
                             handleSearch={handleSearchAction}
                             isMobileSelectionModeEnabled={isMobileSelectionModeEnabled}
                             currentSelectedPolicyID={currentSelectedPolicyID}
                             currentSelectedReportID={currentSelectedReportID}
                             latestBankItems={latestBankItems}
-                            confirmPayment={confirmPayment}
                         />
                     </>
                 )}
