+++ src/pages/Search/SearchPageWide.tsx
@@ -3,7 +3,6 @@ import type {NativeScrollEvent, NativeSyntheticEvent} from 'react-native';
 import {View} from 'react-native';
 import type {OnyxEntry} from 'react-native-onyx';
 import FullPageNotFoundView from '@components/BlockingViews/FullPageNotFoundView';
-import type {DropdownOption} from '@components/ButtonWithDropdownMenu/types';
 import DragAndDropConsumer from '@components/DragAndDrop/Consumer';
 import DragAndDropProvider from '@components/DragAndDrop/Provider';
 import DropZoneUI from '@components/DropZone/DropZoneUI';
@@ -13,29 +12,26 @@ import SearchPageFooter from '@components/Search/SearchPageFooter';
 import SearchFiltersBar from '@components/Search/SearchPageHeader/SearchFiltersBar';
 import SearchPageHeader from '@components/Search/SearchPageHeader/SearchPageHeader';
 import SearchSelectionBar from '@components/Search/SearchPageHeader/SearchSelectionBar';
-import type {SearchHeaderOptionValue} from '@components/Search/SearchPageHeader/SearchPageHeader';
 import type {BankAccountMenuItem, SearchParams, SearchQueryJSON} from '@components/Search/types';
 import {useMemoizedLazyExpensifyIcons} from '@hooks/useLazyAsset';
 import useLocalize from '@hooks/useLocalize';
+import useSearchBulkActions from '@hooks/useSearchBulkActions';
 import useTheme from '@hooks/useTheme';
 import useThemeStyles from '@hooks/useThemeStyles';
 import {buildCannedSearchQuery} from '@libs/SearchQueryUtils';
 import Navigation from '@navigation/Navigation';
 import ROUTES from '@src/ROUTES';
 import type {SearchResults} from '@src/types/onyx';
-import type {PaymentMethodType} from '@src/types/onyx/OriginalMessage';
 
 type SearchPageWideProps = {
     queryJSON?: SearchQueryJSON;
     searchResults: OnyxEntry<SearchResults>;
     searchRequestResponseStatusCode: number | null;
     isMobileSelectionModeEnabled: boolean;
-    headerButtonsOptions: Array<DropdownOption<SearchHeaderOptionValue>>;
     selectedPolicyIDs: Array<string | undefined>;
     selectedTransactionReportIDs: string[];
     selectedReportIDs: string[];
     latestBankItems?: BankAccountMenuItem[];
-    onBulkPaySelected: (paymentMethod?: PaymentMethodType) => void;
     handleSearchAction: (value: SearchParams | string) => void;
     onSortPressedCallback: () => void;
     scrollHandler: (event: NativeSyntheticEvent<NativeScrollEvent>) => void;
@@ -49,12 +45,10 @@ function SearchPageWide({
     searchResults,
     searchRequestResponseStatusCode,
     isMobileSelectionModeEnabled,
-    headerButtonsOptions,
     selectedPolicyIDs,
     selectedTransactionReportIDs,
     selectedReportIDs,
     latestBankItems,
-    onBulkPaySelected,
     handleSearchAction,
     onSortPressedCallback,
     scrollHandler,
@@ -62,6 +56,7 @@ function SearchPageWide({
     PDFValidationComponent,
     ErrorModal,
 }: SearchPageWideProps) {
+    const {headerButtonsOptions} = useSearchBulkActions();
     const styles = useThemeStyles();
     const theme = useTheme();
     const {translate} = useLocalize();
@@ -90,17 +85,14 @@ function SearchPageWide({
                             {PDFValidationComponent}
                             <SearchPageHeader
                                 queryJSON={queryJSON}
-                                headerButtonsOptions={headerButtonsOptions}
                                 handleSearch={handleSearchAction}
                                 isMobileSelectionModeEnabled={isMobileSelectionModeEnabled}
                             />
                             {headerButtonsOptions.length > 0 ? (
                                 <SearchSelectionBar
                                     queryJSON={queryJSON}
-                                    headerButtonsOptions={headerButtonsOptions}
                                     currentSelectedPolicyID={selectedPolicyIDs?.at(0)}
                                     currentSelectedReportID={selectedTransactionReportIDs?.at(0) ?? selectedReportIDs?.at(0)}
-                                    confirmPayment={onBulkPaySelected}
                                     latestBankItems={latestBankItems}
                                 />
                             ) : (
@@ -110,7 +102,6 @@ function SearchPageWide({
                                     isMobileSelectionModeEnabled={isMobileSelectionModeEnabled}
                                     currentSelectedPolicyID={selectedPolicyIDs?.at(0)}
                                     currentSelectedReportID={selectedTransactionReportIDs?.at(0) ?? selectedReportIDs?.at(0)}
-                                    confirmPayment={onBulkPaySelected}
                                     latestBankItems={latestBankItems}
                                 />
                             )}
